fix: implicit declarations, incompatible types, argv type
  fixed a lot of compiler warnings, notably:

chkshsgr.c:10:19: warning: passing argument 2 of ‘getgroups’ from incompatible pointer type [-Wincompatible-pointer-types]
chkshsgr.c:10:32: warning: implicit declaration of function ‘setgroups’; did you mean ‘getgroups’? [-Wimplicit-function-declaration]
prot.c:15:7: warning: implicit declaration of function ‘setgroups’; did you mean ‘getgroups’? [-Wimplicit-function-declaration]
seek_set.c:9:7: warning: implicit declaration of function ‘lseek’ [-Wimplicit-function-declaration]
 - add missing header files, change x type to gid_t

pathexec_env.c:67:22: warning: passing argument 2 of ‘pathexec_run’ from incompatible pointer type [-Wincompatible-pointer-types]
 - change argv and envp types from 'const char * const *argv' to 'char * const *argv'

socklog.c:227:66: warning: pointer targets in passing argument 6 of ‘recvfrom’ differ in signedness [-Wpointer-sign]
 - dummy type should be socklen_t, not int

socklog-check.c:48:3: warning: ‘strncpy’ specified bound 108 equals destination size [-Wstringop-truncation]
socklog.c:171:3: warning: ‘strncpy’ specified bound 108 equals destination size [-Wstringop-truncation]
 - avoid buffer overflow by limiting the copied length

socklog-conf.c:76:19: warning: passing argument 2 of ‘buffer_init’ from incompatible pointer type [-Wincompatible-pointer-types]
buffer.h:18:34: note: expected ‘int (*)()’ but argument is of type ‘ssize_t (*)(int,  const void *, size_t)’ {aka ‘long int (*)(int,  const void *, long unsigned int)’}
 - change buffer_init 2nd parameter from 'int (*)()' to 'ssize_t (*)()'

+++ src/buffer.h
@@ -3,19 +3,21 @@
 #ifndef BUFFER_H
 #define BUFFER_H
 
+#include <sys/types.h>
+
 typedef struct buffer {
   char *x;
   unsigned int p;
   unsigned int n;
   int fd;
-  int (*op)();
+  ssize_t (*op)();
 } buffer;
 
 #define BUFFER_INIT(op,fd,buf,len) { (buf), 0, (len), (fd), (op) }
 #define BUFFER_INSIZE 8192
 #define BUFFER_OUTSIZE 8192
 
-extern void buffer_init(buffer *,int (*)(),int,char *,unsigned int);
+extern void buffer_init(buffer *,ssize_t (*)(),int,char *,unsigned int);
 
 extern int buffer_flush(buffer *);
 extern int buffer_put(buffer *,const char *,unsigned int);
@@ -49,8 +51,8 @@ extern void buffer_seek(buffer *,unsigned int);
 
 extern int buffer_copy(buffer *,buffer *);
 
-extern int buffer_unixread(int,char *,unsigned int);
-extern int buffer_unixwrite(int,const char *,unsigned int);
+extern ssize_t buffer_unixread(int,char *,unsigned int);
+extern ssize_t buffer_unixwrite(int,const char *,unsigned int);
 
 extern buffer *buffer_0;
 extern buffer *buffer_0small;
