+++ src/buffer_write.c
@@ -3,7 +3,7 @@
 #include <unistd.h>
 #include "buffer.h"
 
-int buffer_unixwrite(int fd,const char *buf,unsigned int len)
+ssize_t buffer_unixwrite(int fd,const char *buf,unsigned int len)
 {
   return write(fd,buf,len);
 }
